Nouveau dossier/requirements.txt
MetaTrader5>=5.0.45
pandas>=2.2.0
numpy>=1.26.4
schedule==1.2.1
python-dotenv==1.0.0
streamlit>=1.31.0
plotly==5.18.0
rich==13.7.0
colorlog==6.8.0
google-generativeai==0.3.2
openai==1.12.0
matplotlib>=3.10.5
scipy>=1.12.0
scikit-learn>=1.4.0
requests==2.31.0
# War Room Dashboard
dash>=2.16.0
dash-bootstrap-components>=1.5.0
